passing data from the Character Planner selection bar to the selection of Disciplines
